requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.4.0=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.13.0=pypi_0
astunparse=1.6.3=pypi_0
beautifulsoup4=4.14.3=pypi_0
bertopic=0.17.4=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2026.3.19=hca03da5_0
certifi=2026.2.25=pypi_0
charset-normalizer=3.4.7=pypi_0
click=8.3.2=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
exceptiongroup=1.3.1=pypi_0
filelock=3.25.2=pypi_0
flatbuffers=25.12.19=pypi_0
fonttools=4.62.1=pypi_0
fsspec=2026.3.0=pypi_0
gast=0.7.0=pypi_0
gdown=6.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.80.0=pypi_0
h11=0.16.0=pypi_0
h5py=3.14.0=pypi_0
hdbscan=0.8.42=pypi_0
hf-xet=1.4.3=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=1.10.2=pypi_0
idna=3.11=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.3=pypi_0
keras=3.12.1=pypi_0
kiwisolver=1.5.0=pypi_0
libclang=18.1.1=pypi_0
libcxx=22.1.2=hf894667_0
libexpat=2.7.5=h50f4ffc_0
libffi=3.4.4=hca03da5_1
libzlib=1.3.1=h5f15de7_0
llvmlite=0.47.0=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
matplotlib=3.10.8=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.5.4=pypi_0
mpmath=1.3.0=pypi_0
namex=0.1.0=pypi_0
narwhals=2.20.0=pypi_0
ncurses=6.5=hee39554_0
networkx=3.4.2=pypi_0
numba=0.65.0=pypi_0
numpy=2.2.6=pypi_0
ollama=0.6.1=pypi_0
opencv-python-headless=4.13.0.92=pypi_0
openssl=3.5.6=ha0b305a_0
opt-einsum=3.4.0=pypi_0
optree=0.19.0=pypi_0
packaging=26.0=py310hca03da5_0
pandas=2.3.3=pypi_0
pillow=12.2.0=pypi_0
pip=26.0.1=pyhc872135_1
plotly=6.7.0=pypi_0
protobuf=7.34.1=pypi_0
pydantic=2.13.3=pypi_0
pydantic-core=2.46.3=pypi_0
pygments=2.20.0=pypi_0
pynndescent=0.6.0=pypi_0
pyparsing=3.3.2=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.20=h4f1bc5c_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2026.1.post1=pypi_0
pyyaml=6.0.3=pypi_0
readline=8.3=h0b18652_0
regex=2026.4.4=pypi_0
requests=2.33.1=pypi_0
rich=15.0.0=pypi_0
safetensors=0.7.0=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=pypi_0
sentence-transformers=5.4.1=pypi_0
setuptools=81.0.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
soupsieve=2.8.3=pypi_0
sqlite=3.51.2=h67002bf_0
sympy=1.14.0=pypi_0
tensorflow=2.21.0=pypi_0
termcolor=3.3.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=hcd8a7d5_0
tokenizers=0.22.2=pypi_0
torch=2.11.0=pypi_0
torchvision=0.26.0=pypi_0
tqdm=4.67.3=pypi_0
transformers=5.5.4=pypi_0
typer=0.24.1=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspection=0.4.2=pypi_0
tzdata=2026.1=pypi_0
umap-learn=0.5.12=pypi_0
urllib3=2.6.3=pypi_0
wheel=0.46.3=py310hca03da5_0
wrapt=2.1.2=pypi_0
xz=5.8.2=h8bbcb1d_0
zlib=1.3.1=h5f15de7_0
